test(grid): refactor MCH tests #8119
